o updated the cookie storage dir location o fixed an NPE occurred while deleting a cookie
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/apacheds-replication@755698 13f79535-47bb-0310-9956-ffa450edef68

+++ mitosis/src/main/java/org/apache/directory/mitosis/syncrepl/SyncreplConsumer.java
@@ -150,7 +150,7 @@ public class SyncreplConsumer implements ConsumerCallback
     {
         this.directoryService = directoryservice;
 
-        File cookieDir = new File( "cookies" );
+        File cookieDir = new File( directoryservice.getWorkingDirectory(), "cookies" );
         cookieDir.mkdir();
         
         cookieFile = new File( cookieDir, String.valueOf( config.getReplicaId() ) );
@@ -671,7 +671,7 @@ public class SyncreplConsumer implements ConsumerCallback
      */
     public void deleteCookieFile()
     {
-        if( cookieFile.exists() )
+        if( cookieFile != null && cookieFile.exists() )
         {
             LOG.debug( "deleting the cookie file" );
             cookieFile.delete();
